+++ src/cli/lib.ts
@@ -1,6 +1,6 @@
 import fs from 'fs'
 import { CarIndexedReader, CarCIDIterator } from '@ipld/car'
-import { fromCar } from '../from-car'
+import { fromCar } from '../unpack'
 
 export async function listFilesInCar ({input}: {input: string}) {
   const carReader = await CarIndexedReader.fromFile(input)
