+++ src/pack/fs.ts
@@ -1,0 +1,20 @@
+import fs from 'fs'
+import os from 'os'
+import path from 'path'
+
+import { toCar } from '.'
+import { Blockstore } from '../blockstore'
+import { FsBlockStore } from '../blockstore/fs'
+
+export default async function packToFs ({ input, output, blockstore = new FsBlockStore() }: { input: string | Iterable<string> | AsyncIterable<string>, output?: string, blockstore?: Blockstore }) {
+  const location = output || `${os.tmpdir()}/${(parseInt(String(Math.random() * 1e9), 10)).toString() + Date.now()}`
+  const writable = fs.createWriteStream(location)
+
+  const root = await toCar({ input, writable, blockstore })
+
+  // Move to work dir
+  if (!output) {
+    const inputName = typeof input === 'string' ? path.parse(path.basename(input)).name : root.toString()
+    await fs.promises.rename(location, `${process.cwd()}/${inputName}.car`)
+  }
+}
