+++ test/pack/index.node.test.ts
@@ -7,46 +7,18 @@ const equals = require('uint8arrays/equals')
 
 import { CarReader } from '@ipld/car'
 
-import { fromCar } from '../dist/from-car'
+import { fromCar } from '../../dist/from-car'
 
-import {
-  packFileIterableToCar,
-  packFileToCarFs,
-  toCar
-} from '../dist/to-car'
+import pack from '../../dist/pack'
+import packToFs from '../../dist/pack/fs'
 
-import { MemoryBlockStore } from '../dist/blockstore/memory'
-import { FsBlockStore } from '../dist/blockstore/fs'
-import { LevelBlockStore } from '../dist/blockstore/level'
+import { MemoryBlockStore } from '../../dist/blockstore/memory'
+import { FsBlockStore } from '../../dist/blockstore/fs'
+import { LevelBlockStore } from '../../dist/blockstore/level'
 
 const dirTmp = `${__dirname}/tmp`
 
-describe('toCar', () => {
-  describe('toCar basic', () => {
-    beforeEach(() => {
-      if (!fs.existsSync(dirTmp)) {
-        fs.mkdirSync(dirTmp)
-      }
-    })
-
-    afterEach(() => {
-      fs.rmdirSync(dirTmp, { recursive: true })
-    })
-
-    it('read a raw car file and transform it back toCar', async () => {
-      const inStream = fs.createReadStream(`${__dirname}/fixtures/raw.car`)
-      const carReader = await CarReader.fromIterable(inStream)
-      const files = await all(fromCar(carReader))
-
-      expect(files).to.have.lengthOf(1)
-
-      await toCar({
-        files,
-        writable: fs.createWriteStream(`${__dirname}/tmp/raw.car`)
-      })
-    })
-  })
-
+describe('pack', () => {
   ;[MemoryBlockStore, FsBlockStore, LevelBlockStore].map((Blockstore) => {
     describe(`with ${Blockstore.name}`, () => {
       beforeEach(() => {
@@ -68,8 +40,8 @@ describe('toCar', () => {
       it('pack dir to car with filesystem output with iterable input', async () => {
         const writable = fs.createWriteStream(`${__dirname}/tmp/dir.car`)
         // Create car from file
-        await packFileIterableToCar({
-          input: `${__dirname}/fixtures/dir`,
+        await pack({
+          input: `${__dirname}/../fixtures/dir`,
           writable,
           blockstore: new Blockstore()
         })
@@ -83,8 +55,8 @@ describe('toCar', () => {
 
       it('pack dir to car with filesystem output', async () => {
         // Create car from file
-        await packFileToCarFs({
-          input: `${__dirname}/fixtures/dir`,
+        await packToFs({
+          input: `${__dirname}/../fixtures/dir`,
           output: `${__dirname}/tmp/dir.car`,
           blockstore: new Blockstore()
         })
@@ -98,8 +70,8 @@ describe('toCar', () => {
 
       it('pack raw file to car with filesystem output', async () => {
         // Create car from file
-        await packFileToCarFs({
-          input: `${__dirname}/fixtures/file.raw`,
+        await packToFs({
+          input: `${__dirname}/../fixtures/file.raw`,
           output: `${__dirname}/tmp/raw.car`,
           blockstore: new Blockstore()
         })
@@ -110,7 +82,7 @@ describe('toCar', () => {
 
         expect(files).to.have.lengthOf(1)
 
-        const rawOriginalContent = new Uint8Array(fs.readFileSync(`${__dirname}/fixtures/file.raw`))
+        const rawOriginalContent = new Uint8Array(fs.readFileSync(`${__dirname}/../fixtures/file.raw`))
         const rawContent = (await all(files[0].content()))[0]
 
         expect(equals(rawOriginalContent, rawContent)).to.eql(true)
@@ -118,8 +90,8 @@ describe('toCar', () => {
 
       it('pack raw file to car without output', async () => {
         // Create car from file
-        await packFileToCarFs({
-          input: `${__dirname}/fixtures/file.raw`,
+        await packToFs({
+          input: `${__dirname}/../fixtures/file.raw`,
           blockstore: new Blockstore()
         })
         const newCarPath = `${process.cwd()}/file.car`
@@ -130,7 +102,7 @@ describe('toCar', () => {
 
         expect(files).to.have.lengthOf(1)
 
-        const rawOriginalContent = new Uint8Array(fs.readFileSync(`${__dirname}/fixtures/file.raw`))
+        const rawOriginalContent = new Uint8Array(fs.readFileSync(`${__dirname}/../fixtures/file.raw`))
         const rawContent = (await all(files[0].content()))[0]
 
         expect(equals(rawOriginalContent, rawContent)).to.eql(true)
@@ -143,8 +115,8 @@ describe('toCar', () => {
         const writable = fs.createWriteStream(`${__dirname}/tmp/raw.car`)
 
         // Create car from file
-        await packFileIterableToCar({
-          input: `${__dirname}/fixtures/file.raw`,
+        await pack({
+          input: `${__dirname}/../fixtures/file.raw`,
           writable,
           blockstore: new Blockstore()
         })
@@ -155,7 +127,7 @@ describe('toCar', () => {
 
         expect(files).to.have.lengthOf(1)
 
-        const rawOriginalContent = new Uint8Array(fs.readFileSync(`${__dirname}/fixtures/file.raw`))
+        const rawOriginalContent = new Uint8Array(fs.readFileSync(`${__dirname}/../fixtures/file.raw`))
         const rawContent = (await all(files[0].content()))[0]
 
         expect(equals(rawOriginalContent, rawContent)).to.eql(true)
